requirements.txt
python-dateutil
parameterized
